+++ indy_common/test/auth/metadata/conftest.py
@@ -4,12 +4,11 @@ import pytest
 
 from indy_common.authorize.auth_constraints import IDENTITY_OWNER
 from indy_common.constants import TRUST_ANCHOR
-from indy_common.test.auth.conftest import OTHER_ROLE
 from indy_common.test.auth.metadata.helper import PluginAuthorizer
 from plenum.common.constants import TRUSTEE, STEWARD
 from plenum.test.conftest import getValueFromModule
 
-ROLES = [TRUSTEE, STEWARD, TRUST_ANCHOR, IDENTITY_OWNER, OTHER_ROLE]
+ROLES = [TRUSTEE, STEWARD, TRUST_ANCHOR, IDENTITY_OWNER]
 MAX_SIG_COUNT = 3
